+++ nautilus_trader/adapters/binance/websocket/spot.py
@@ -163,9 +163,3 @@ class BinanceSpotWebSocket(BinanceWebSocketClient):
 
         """
         self._add_stream(f"{format_symbol(symbol)}@depth@{speed}ms")
-
-    def subscribe_user_data(self, listen_key: str):
-        """
-        Listen to user data by provided `listenkey`.
-        """
-        self._add_stream(listen_key)