+++ nautilus_trader/common/events/risk.pyx
@@ -78,8 +78,10 @@ cdef class TradingStateChanged(RiskEvent):
         ----------
         trader_id : TraderId
             The trader ID associated with the event.
-        trader_id : TraderId
+        state : TradingState
             The trading state for the event.
+        config : dict
+            The configuration of the risk engine.
         event_id : UUID4
             The event ID.
         ts_event : int64