+++ nautilus_trader/examples/strategies/ema_cross.py
@@ -114,14 +114,14 @@ class EMACross(TradingStrategy):
         self.register_indicator_for_bars(self.bar_type, self.slow_ema)
 
         # Get historical data
-        # self.request_bars(self.bar_type)
+        self.request_bars(self.bar_type)
         # self.request_quote_ticks(self.instrument_id)
         # self.request_trade_ticks(self.instrument_id)
 
         # Subscribe to live data
-        self.subscribe_bars(self.bar_type)  # For debugging
+        self.subscribe_bars(self.bar_type)
+        self.subscribe_quote_ticks(self.instrument_id)
         # self.subscribe_ticker(self.instrument_id)  # For debugging
-        # self.subscribe_quote_ticks(self.instrument_id)  # For debugging
         # self.subscribe_trade_ticks(self.instrument_id)  # For debugging
         # self.subscribe_order_book_deltas(self.instrument_id, depth=20)  # For debugging
         # self.subscribe_order_book_snapshots(self.instrument_id, depth=20)  # For debugging
@@ -187,7 +187,7 @@ class EMACross(TradingStrategy):
             The quote tick received.
 
         """
-        self.log.info(f"Received {repr(tick)}")  # For debugging (must add a subscription)
+        # self.log.info(f"Received {repr(tick)}")  # For debugging (must add a subscription)
 
     def on_trade_tick(self, tick: TradeTick):
         """
@@ -295,11 +295,12 @@ class EMACross(TradingStrategy):
         self.flatten_all_positions(self.instrument_id)
 
         # Unsubscribe from data
+        self.unsubscribe_bars(self.bar_type)
+        self.unsubscribe_quote_ticks(self.instrument_id)
+
         # self.unsubscribe_ticker(self.instrument_id)
         # self.unsubscribe_order_book_deltas(self.instrument_id)
-        self.unsubscribe_bars(self.bar_type)
         # self.unsubscribe_order_book_snapshots(self.instrument_id)
-        # self.unsubscribe_quote_ticks(self.instrument_id)
         # self.unsubscribe_trade_ticks(self.instrument_id)
 
     def on_reset(self):