+++ tests/integration_tests/adapters/binance/test_providers.py
@@ -16,6 +16,7 @@
 import pkgutil
 from typing import Dict
 
+import orjson
 import pytest
 
 from nautilus_trader.adapters.binance.http.client import BinanceHttpClient
@@ -53,7 +54,7 @@ class TestBinanceInstrumentProvider:
             url_path: str,  # noqa (needed for mock)
             payload: Dict[str, str],  # noqa (needed for mock)
         ) -> bytes:
-            return responses.pop()
+            return orjson.loads(responses.pop())
 
         # Apply mock coroutine to client
         monkeypatch.setattr(